Calender.php
This is the only and main PHP file in which i've give some comment notes for better understanding. i think you are a php programmer and know a little-bit of JQuery then i then you can understand my whole program. though i've colaborate html,CSS, JQuery and PHP in a single file but it will give you a clear and brief understanding of it.